requirements.txt
opencv-python-headless
ultralytics
face_recognition
firebase-admin
google-genai
requests
Pillow
fastapi
uvicorn
